@@ -5,7 +5,6 @@ from __future__ import annotations
 import logging
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 from jobflow import Flow, Job, Maker, OutputReference
@@ -21,9 +20,12 @@ from atomate2.common.jobs.defect import (
 )
 
 if TYPE_CHECKING:
+    from pathlib import Path
+
     import numpy.typing as npt
     from pymatgen.analysis.defects.core import Defect
     from pymatgen.core.structure import Structure
+    from pymatgen.entries.computed_entries import ComputedStructureEntry
 
 logger = logging.getLogger(__name__)
 
@@ -110,7 +112,7 @@ class ConfigurationCoordinateMaker(Maker):
             struct2,
             distortions=self.distortions,
             static_maker=self.static_maker,
-            prev_vasp_dir=dir1,
+            prev_dir=dir1,
             add_name="q1",
             add_info={"relaxed_uuid": relax1.uuid, "distorted_uuid": relax2.uuid},
         )
@@ -120,7 +122,7 @@ class ConfigurationCoordinateMaker(Maker):
             struct1,
             distortions=self.distortions,
             static_maker=self.static_maker,
-            prev_vasp_dir=dir2,
+            prev_dir=dir2,
             add_name="q2",
             add_info={"relaxed_uuid": relax2.uuid, "distorted_uuid": relax1.uuid},
         )
@@ -295,6 +297,7 @@ class FormationEnergyMaker(Maker, ABC):
                 uc_structure=defect.structure,
                 relax_maker=self.bulk_relax_maker,
                 sc_mat=supercell_matrix,
+                get_planar_locpot=self.get_planar_locpot,
             )
             sc_mat = get_sc_job.output["sc_mat"]
             lattice = get_sc_job.output["sc_struct"].lattice
@@ -305,8 +308,8 @@ class FormationEnergyMaker(Maker, ABC):
             get_sc_job = get_supercell_from_prv_calc(
                 uc_structure=defect.structure,
                 prv_calc_dir=bulk_supercell_dir,
+                sc_entry_and_locpot_from_prv=self.sc_entry_and_locpot_from_prv,
                 sc_mat_ref=supercell_matrix,
-                structure_from_prv=self.structure_from_prv,
             )
             sc_mat = get_sc_job.output["sc_mat"]
             lattice = get_sc_job.output["lattice"]
@@ -329,12 +332,6 @@ class FormationEnergyMaker(Maker, ABC):
         jobs.extend([get_sc_job, spawn_output])
 
         if self.collect_defect_entry_data:
-            if isinstance(bulk_supercell_dir, (str, Path)):
-                raise NotImplementedError(
-                    "DefectEntery creation only works when you are explicitly "
-                    "calculating the bulk supercell. This is because the bulk "
-                    "SC energy parsing from previous calculations is not implemented."
-                )
             collection_job = get_defect_entry(
                 charge_state_summary=spawn_output.output,
                 bulk_summary=get_sc_job.output,
@@ -348,8 +345,10 @@ class FormationEnergyMaker(Maker, ABC):
         )
 
     @abstractmethod
-    def structure_from_prv(self, previous_dir: str) -> Structure:
-        """Copy the output structure from previous directory.
+    def sc_entry_and_locpot_from_prv(
+        self, previous_dir: str
+    ) -> tuple[ComputedStructureEntry, dict]:
+        """Copy the output ComputedStructureEntry and Locpot from previous directory.
 
         Parameters
         ----------
@@ -358,7 +357,25 @@ class FormationEnergyMaker(Maker, ABC):
 
         Returns
         -------
-        structure: Structure
+        entry: ComputedStructureEntry
+        """
+
+    @abstractmethod
+    def get_planar_locpot(self, task_doc) -> dict:
+        """Get the Planar Locpot from the TaskDoc.
+
+        This is needed just in case the planar average locpot is stored in different
+        part of the TaskDoc for different codes.
+
+        Parameters
+        ----------
+        task_doc: TaskDoc
+            The task document.
+
+        Returns
+        -------
+        planar_locpot: dict
+            The planar average locpot.
         """
 
     @abstractmethod
