@@ -48,7 +48,8 @@ LOBSTER_UNIFORM_MAKER = UniformBandStructureMaker(
                 "LWAVE": True,
                 "ISYM": 0,
             },
-        )
+        ),
+        task_document_kwargs={"parse_dos": False, "parse_bandstructure": False},
     ),
 )
 
@@ -61,9 +62,8 @@ class VaspLobsterMaker(Maker):
     The calculations performed are:
 
     1. Optional optimization.
-    2. Optional static computation with symmetry to preconverge the wavefunction.
-    3. Static calculation with ISYM=0.
-    4. Several Lobster computations testing several basis sets are performed.
+    2. Static calculation with ISYM=0.
+    3. Several Lobster computations testing several basis sets are performed.
 
     .. Note::
 
@@ -76,9 +76,6 @@ class VaspLobsterMaker(Maker):
     relax_maker : .BaseVaspMaker or None
         A maker to perform a relaxation on the bulk. Set to ``None`` to skip the
         bulk relaxation.
-    preconverge_static_maker : .BaseVaspMaker or None
-        A maker to perform a preconvergence run before the wavefunction computation
-        without symmetry
     lobster_static_maker : .BaseVaspMaker
         A maker to perform the computation of the wavefunction before the static run.
         Cannot be skipped. It can be LOBSTERUNIFORM or LobsterStaticMaker()
@@ -107,7 +104,7 @@ class VaspLobsterMaker(Maker):
     def make(
         self,
         structure: Structure,
-        prev_vasp_dir: str | Path | None = None,
+        prev_dir: str | Path | None = None,
     ) -> Flow:
         """
         Make flow to calculate bonding properties.
@@ -118,7 +115,7 @@ class VaspLobsterMaker(Maker):
             A pymatgen structure. Please start with a structure
             that is nearly fully optimized as the internal optimizers
             have very strict settings!
-        prev_vasp_dir : str or Path or None
+        prev_dir : str or Path or None
             A previous vasp calculation directory to use for copying outputs.
         """
         jobs = []
@@ -127,12 +124,12 @@ class VaspLobsterMaker(Maker):
         optimization_dir = None
         optimization_uuid = None
         if self.relax_maker is not None:
-            optimization = self.relax_maker.make(structure, prev_vasp_dir=prev_vasp_dir)
+            optimization = self.relax_maker.make(structure, prev_dir=prev_dir)
             jobs.append(optimization)
             structure = optimization.output.structure
             optimization_dir = optimization.output.dir_name
             optimization_uuid = optimization.output.uuid
-            prev_vasp_dir = optimization_dir
+            prev_dir = optimization_dir
 
         # Information about the basis is collected
         basis_infos = get_basis_infos(
@@ -149,7 +146,7 @@ class VaspLobsterMaker(Maker):
             self.lobster_static_maker,
             basis_infos.output["nbands"],
             structure,
-            prev_vasp_dir,
+            prev_dir,
         )
         jobs.append(lobster_static)
         lobster_static_dir = lobster_static.output.dir_name
