@@ -34,7 +34,6 @@ class MPGGARelaxSetGenerator(RelaxSetGenerator):
 
     config_dict: dict = field(default_factory=lambda: _BASE_MP_GGA_RELAX_SET)
     auto_kspacing: bool = True
-    inherit_incar: bool = False
 
 
 @dataclass
@@ -43,7 +42,6 @@ class MPGGAStaticSetGenerator(StaticSetGenerator):
 
     config_dict: dict = field(default_factory=lambda: _BASE_MP_GGA_RELAX_SET)
     auto_kspacing: bool = True
-    inherit_incar: bool = False
 
     def get_incar_updates(
         self,
@@ -90,7 +88,7 @@ class MPMetaGGAStaticSetGenerator(StaticSetGenerator):
 
     config_dict: dict = field(default_factory=lambda: _BASE_MP_R2SCAN_RELAX_SET)
     auto_kspacing: bool = True
-    inherit_incar: bool = False
+    bandgap_tol: float = 1e-4
 
     def get_incar_updates(
         self,
@@ -148,7 +146,6 @@ class MPMetaGGARelaxSetGenerator(RelaxSetGenerator):
     config_dict: dict = field(default_factory=lambda: _BASE_MP_R2SCAN_RELAX_SET)
     bandgap_tol: float = 1e-4
     auto_kspacing: bool = True
-    inherit_incar: bool = False
 
     def get_incar_updates(
         self,
