@@ -1,11 +1,24 @@
+import logging
+import os
+import shutil
+import sys
+import tempfile
+from pathlib import Path
+from unittest import mock
+
 import pytest
+from fireworks import LaunchPad
+from jobflow import JobStore
+from jobflow.settings import JobflowSettings
+from maggma.stores import MemoryStore
 from monty.serialization import loadfn
+from pymatgen.core import Structure
+
+from atomate2.utils.log import initialize_logger
 
 
 @pytest.fixture(scope="session")
 def test_dir():
-    from pathlib import Path
-
     module_dir = Path(__file__).resolve().parent
     test_dir = module_dir / "test_data"
     return test_dir.resolve()
@@ -13,9 +26,6 @@ def test_dir():
 
 @pytest.fixture(scope="session")
 def log_to_stdout():
-    import logging
-    import sys
-
     # Set Logging
     root = logging.getLogger()
     root.setLevel(logging.DEBUG)
@@ -29,10 +39,6 @@ def log_to_stdout():
 
 @pytest.fixture(scope="session")
 def clean_dir(debug_mode):
-    import os
-    import shutil
-    import tempfile
-
     old_cwd = os.getcwd()
     new_path = tempfile.mkdtemp()
     os.chdir(new_path)
@@ -47,9 +53,6 @@ def clean_dir(debug_mode):
 @pytest.fixture()
 def tmp_dir():
     """Same as clean_dir but is fresh for every test"""
-    import os
-    import shutil
-    import tempfile
 
     old_cwd = os.getcwd()
     new_path = tempfile.mkdtemp()
@@ -66,8 +69,6 @@ def debug_mode():
 
 @pytest.fixture(scope="session")
 def lpad(database, debug_mode):
-    from fireworks import LaunchPad
-
     lpad = LaunchPad(name=database)
     lpad.reset("", require_password=False)
     yield lpad
@@ -80,9 +81,6 @@ def lpad(database, debug_mode):
 
 @pytest.fixture()
 def memory_jobstore():
-    from jobflow import JobStore
-    from maggma.stores import MemoryStore
-
     store = JobStore(MemoryStore(), additional_stores={"data": MemoryStore()})
     store.connect()
 
@@ -91,24 +89,17 @@ def memory_jobstore():
 
 @pytest.fixture(scope="session", autouse=True)
 def log_to_stdout_auto_use():
-    from atomate2.utils.log import initialize_logger
-
     initialize_logger()
 
 
 @pytest.fixture()
 def si_structure(test_dir):
-    from pymatgen.core import Structure
-
     return Structure.from_file(test_dir / "structures" / "Si.cif")
 
 
 @pytest.fixture(autouse=True)
 def mock_jobflow_settings(memory_jobstore):
     """Mock the jobflow settings to use our specific jobstore (with data store)."""
-    from unittest import mock
-
-    from jobflow.settings import JobflowSettings
 
     settings = JobflowSettings(JOB_STORE=memory_jobstore)
 
