@@ -1,4 +1,5 @@
-import numpy as np
+from jobflow import run_locally
+from numpy.testing import assert_allclose
 from pymatgen.core.structure import Structure
 from pymatgen.phonon.bandstructure import PhononBandStructureSymmLine
 from pymatgen.phonon.dos import PhononDos
@@ -13,15 +14,12 @@ from atomate2.forcefields.flows.phonons import PhononMaker
 
 
 def test_phonon_wf(clean_dir):
-    from jobflow import run_locally
-
     structure = Structure(
         lattice=[[0, 2.73, 2.73], [2.73, 0, 2.73], [2.73, 2.73, 0]],
         species=["Si", "Si"],
         coords=[[0, 0, 0], [0.25, 0.25, 0.25]],
     )
 
-    # !!! Generate job
     job = PhononMaker(
         use_symmetrized_structure="conventional",
         create_thermal_displacements=False,
@@ -33,22 +31,13 @@ def test_phonon_wf(clean_dir):
     # run the flow or job and ensure that it finished running successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
-    # !!! validation on the outputs
+    # validate the outputs
     assert isinstance(responses[job.jobs[-1].uuid][1].output, PhononBSDOSDoc)
 
-    assert np.allclose(
-        np.array(responses[job.jobs[-1].uuid][1].output.free_energies) / 1000.0,
-        np.array(
-            [
-                5058.45217527524,
-                4907.495751683517,
-                3966.5493299635937,
-                2157.8178928940474,
-                -357.5054580420707,
-            ]
-        )
-        / 1000.0,
-        2,
+    assert_allclose(
+        responses[job.jobs[-1].uuid][1].output.free_energies,
+        [5058.4521752, 4907.4957516, 3966.5493299, 2157.8178928, -357.5054580],
+        atol=1000,
     )
 
     assert isinstance(
@@ -58,28 +47,25 @@ def test_phonon_wf(clean_dir):
     assert isinstance(responses[job.jobs[-1].uuid][1].output.phonon_dos, PhononDos)
     assert responses[job.jobs[-1].uuid][1].output.thermal_displacement_data is None
     assert isinstance(responses[job.jobs[-1].uuid][1].output.structure, Structure)
-    assert np.allclose(
+    assert_allclose(
         responses[job.jobs[-1].uuid][1].output.temperatures, [0, 100, 200, 300, 400]
     )
     assert responses[job.jobs[-1].uuid][1].output.force_constants is None
     assert isinstance(responses[job.jobs[-1].uuid][1].output.jobdirs, PhononJobDirs)
     assert isinstance(responses[job.jobs[-1].uuid][1].output.uuids, PhononUUIDs)
-    assert np.isclose(
-        responses[job.jobs[-1].uuid][1].output.total_dft_energy, -5.372457981109619, 4
+    assert_allclose(
+        responses[job.jobs[-1].uuid][1].output.total_dft_energy, -5.37245798, 4
     )
     assert responses[job.jobs[-1].uuid][1].output.born is None
     assert responses[job.jobs[-1].uuid][1].output.epsilon_static is None
-    assert np.allclose(
+    assert_allclose(
         responses[job.jobs[-1].uuid][1].output.supercell_matrix,
         [[4.0, 0.0, 0.0], [0.0, 4.0, 0.0], [0.0, 0.0, 4.0]],
     )
-    assert np.allclose(
+    assert_allclose(
         responses[job.jobs[-1].uuid][1].output.primitive_matrix,
-        (
-            (0, 0.5000000000000001, 0.5000000000000001),
-            (0.5000000000000001, 0.0, 0.5000000000000001),
-            (0.5000000000000001, 0.5000000000000001, 0.0),
-        ),
+        ((0, 0.5, 0.5), (0.5, 0.0, 0.5), (0.5, 0.5, 0.0)),
+        atol=1e-8,
     )
     assert responses[job.jobs[-1].uuid][1].output.code == "vasp"
     assert isinstance(
@@ -95,40 +81,18 @@ def test_phonon_wf(clean_dir):
         responses[job.jobs[-1].uuid][1].output.phonopy_settings.kpoint_density_dos
         == 7000
     )
-    assert np.allclose(
+    assert_allclose(
         responses[job.jobs[-1].uuid][1].output.entropies,
-        [
-            0.0,
-            4.783939817386235,
-            13.993186953791708,
-            21.88641334781562,
-            28.19110667148253,
-        ],
-        3,
+        [0.0, 4.78393981, 13.99318695, 21.88641334, 28.19110667],
+        atol=2,
     )
-    assert np.allclose(
+    assert_allclose(
         responses[job.jobs[-1].uuid][1].output.heat_capacities,
-        [
-            0.0,
-            8.860605865667427,
-            17.55758943495313,
-            21.089039169564796,
-            22.625872713428905,
-        ],
-        3,
+        [0.0, 8.86060586, 17.55758943, 21.08903916, 22.62587271],
+        atol=2,
     )
-
-    assert np.allclose(
-        np.array(responses[job.jobs[-1].uuid][1].output.internal_energies) / 1000.0,
-        np.array(
-            [
-                5058.441587914012,
-                5385.880585798466,
-                6765.198541655172,
-                8723.78588089732,
-                10919.019940938391,
-            ]
-        )
-        / 1000.0,
-        3,
+    assert_allclose(
+        responses[job.jobs[-1].uuid][1].output.internal_energies,
+        [5058.44158791, 5385.88058579, 6765.19854165, 8723.78588089, 10919.0199409],
+        atol=1000,
     )
