@@ -1,14 +1,20 @@
+from jobflow import run_locally
 from pytest import approx, importorskip
+
+from atomate2.forcefields.jobs import (
+    CHGNetRelaxMaker,
+    CHGNetStaticMaker,
+    GAPRelaxMaker,
+    GAPStaticMaker,
+    M3GNetRelaxMaker,
+    M3GNetStaticMaker,
+)
+from atomate2.forcefields.schemas import ForceFieldTaskDocument
 
 importorskip("quippy")
 
 
 def test_chgnet_static_maker(si_structure):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import CHGNetStaticMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
 
     # generate job
@@ -20,17 +26,12 @@ def test_chgnet_static_maker(si_structure):
     # validate job outputs
     output1 = responses[job.uuid][1].output
     assert isinstance(output1, ForceFieldTaskDocument)
-    assert output1.output.energy == approx(-10.7452, rel=1e-4)
+    assert output1.output.energy == approx(-10.6275062, rel=1e-4)
     assert output1.output.ionic_steps[-1].magmoms is None
     assert output1.output.n_steps == 1
 
 
 def test_chgnet_relax_maker(si_structure):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import CHGNetRelaxMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     # translate one atom to ensure a small number of relaxation steps are taken
     si_structure.translate_sites(0, [0, 0, 0.1])
 
@@ -43,17 +44,12 @@ def test_chgnet_relax_maker(si_structure):
     # validate job outputs
     output1 = responses[job.uuid][1].output
     assert isinstance(output1, ForceFieldTaskDocument)
-    assert output1.output.energy == approx(-10.74523544, rel=1e-4)
-    assert output1.output.ionic_steps[-1].magmoms[0] == approx(0.00211287, rel=1e-4)
-    assert output1.output.n_steps == 12
+    assert output1.output.energy == approx(-10.6274, rel=1e-4)
+    assert output1.output.ionic_steps[-1].magmoms[0] == approx(0.00303572, rel=1e-4)
+    assert output1.output.n_steps >= 12
 
 
 def test_m3gnet_static_maker(si_structure):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import M3GNetStaticMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
 
     # generate job
@@ -70,11 +66,6 @@ def test_m3gnet_static_maker(si_structure):
 
 
 def test_m3gnet_relax_maker(si_structure):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import M3GNetRelaxMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     # translate one atom to ensure a small number of relaxation steps are taken
     si_structure.translate_sites(0, [0, 0, 0.1])
 
@@ -92,11 +83,6 @@ def test_m3gnet_relax_maker(si_structure):
 
 
 def test_gap_static_maker(si_structure, test_dir):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import GAPStaticMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     task_doc_kwargs = {"ionic_step_data": ("structure", "energy")}
 
     # generate job
@@ -120,11 +106,6 @@ def test_gap_static_maker(si_structure, test_dir):
 
 
 def test_gap_relax_maker(si_structure, test_dir):
-    from jobflow import run_locally
-
-    from atomate2.forcefields.jobs import GAPRelaxMaker
-    from atomate2.forcefields.schemas import ForceFieldTaskDocument
-
     # translate one atom to ensure a small number of relaxation steps are taken
     si_structure.translate_sites(0, [0, 0, 0.1])
 
