@@ -1,16 +1,24 @@
 from typing import TYPE_CHECKING
 
+import numpy as np
+from jobflow import JobStore, run_locally
+from maggma.stores.mongolike import MemoryStore
+from pymatgen.analysis.defects.generators import SubstitutionGenerator
+from pymatgen.core import Structure
+from pymatgen.io.vasp.outputs import WSWQ
+
+from atomate2.vasp.flows.defect import (
+    ConfigurationCoordinateMaker,
+    FormationEnergyMaker,
+    NonRadiativeMaker,
+)
+
 if TYPE_CHECKING:
     from atomate2.common.schemas.defects import CCDDocument
     from atomate2.vasp.schemas.defect import FiniteDifferenceDocument
 
 
 def test_ccd_maker(mock_vasp, clean_dir, test_dir):
-    from jobflow import run_locally
-    from pymatgen.core import Structure
-
-    from atomate2.vasp.flows.defect import ConfigurationCoordinateMaker
-
     # mapping from job name to directory containing test files
     # mapping from job name to directory containing test files
     ref_paths = {
@@ -57,17 +65,6 @@ def test_ccd_maker(mock_vasp, clean_dir, test_dir):
 
 
 def test_nonrad_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
-    import numpy as np
-    from jobflow import JobStore, run_locally
-    from maggma.stores.mongolike import MemoryStore
-    from pymatgen.core import Structure
-    from pymatgen.io.vasp.outputs import WSWQ
-
-    from atomate2.vasp.flows.defect import (
-        ConfigurationCoordinateMaker,
-        NonRadiativeMaker,
-    )
-
     # mapping from job name to directory containing test files
     ref_paths = {
         "relax q1": "Si_config_coord/relax_q1",
@@ -132,10 +129,6 @@ def test_nonrad_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
 
 def test_formation_energy_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
     from jobflow import SETTINGS, run_locally
-    from pymatgen.analysis.defects.generators import SubstitutionGenerator
-    from pymatgen.core import Structure
-
-    from atomate2.vasp.flows.defect import FormationEnergyMaker
 
     # mapping from job name to directory containing test files
     ref_paths = {
@@ -160,7 +153,6 @@ def test_formation_energy_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
         )
     )
 
-    # rmaker = RelaxMaker(input_set_generator=ChargeStateRelaxSetGenerator())
     maker = FormationEnergyMaker(
         relax_radius="auto",
         perturb=0.1,
@@ -187,3 +179,16 @@ def test_formation_energy_maker(mock_vasp, clean_dir, test_dir, monkeypatch):
 
     for k in ref_paths:
         _check_plnr_locpot(k)
+
+    # make sure the the you can restart the calculation from prv
+    prv_dir = test_dir / "vasp/GaN_Mg_defect/bulk_relax/outputs"
+    flow2 = maker.make(
+        defects[0],
+        bulk_supercell_dir=prv_dir,
+        defect_index=0,
+    )
+    _ = run_locally(
+        flow2,
+        create_folders=True,
+        ensure_success=True,
+    )
