@@ -26,11 +26,23 @@ def test_matpes_gga_plus_meta_gga_static_maker(mock_vasp, clean_dir, vasp_test_d
     assert len(flow) == 2
     assert [job.name for job in flow] == list(ref_paths)
 
+    # make sure first static has LWAVE=True so we can pass its WAVECAR to second static
+    # as pre-conditioned starting point
+    assert flow[0].maker.input_set_generator.user_incar_settings["LWAVE"] is True
+
     # ensure flow runs successfully
     responses = run_locally(flow, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[flow.jobs[-1].uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-17.53895666)
-    assert output.output.bandgap == pytest.approx(0.8087999)
+    pbe_doc = responses[flow.jobs[0].uuid][1].output
+    assert isinstance(pbe_doc, TaskDoc)
+    assert pbe_doc.output.energy == pytest.approx(-10.84940729)
+    assert pbe_doc.output.bandgap == pytest.approx(0.6172, abs=1e-3)
+
+    r2scan_doc = responses[flow.jobs[-1].uuid][1].output
+    assert isinstance(r2scan_doc, TaskDoc)
+    assert r2scan_doc.output.energy == pytest.approx(-17.53895666)
+    assert r2scan_doc.output.bandgap == pytest.approx(0.8087999)
+
+    assert isinstance(flow.output, dict)
+    assert {*flow.output} == {"static1", "static2"}
