@@ -1,18 +1,20 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
 import pytest
+from emmet.core.tasks import TaskDoc
+from jobflow import Maker, run_locally
 from pymatgen.core import Structure
 
+from atomate2.vasp.flows.mp import (
+    MPGGADoubleRelaxMaker,
+    MPGGADoubleRelaxStaticMaker,
+    MPMetaGGADoubleRelaxStaticMaker,
+)
 from atomate2.vasp.jobs.mp import (
     MPMetaGGARelaxMaker,
     MPPreRelaxMaker,
 )
 from atomate2.vasp.sets.mp import MPMetaGGARelaxSetGenerator
-
-if TYPE_CHECKING:
-    from jobflow import Maker
 
 
 @pytest.mark.parametrize("name", ["test", None])
@@ -28,8 +30,6 @@ if TYPE_CHECKING:
 def test_mp_meta_gga_relax_custom_values(
     name: str, relax_maker: Maker | None, static_maker: Maker | None
 ):
-    from atomate2.vasp.flows.mp import MPMetaGGADoubleRelaxStaticMaker
-
     kwargs = {}
     if name:
         kwargs["name"] = name
@@ -48,11 +48,6 @@ def test_mp_meta_gga_relax_custom_values(
 
 
 def test_mp_meta_gga_double_relax_static(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
-    from atomate2.vasp.flows.mp import MPMetaGGADoubleRelaxStaticMaker
-
     # map from job name to directory containing reference output files
     pre_relax_dir = "Si_mp_meta_gga_relax/pbesol_pre_relax"
     ref_paths = {
@@ -77,17 +72,12 @@ def test_mp_meta_gga_double_relax_static(mock_vasp, clean_dir, vasp_test_dir):
     responses = run_locally(flow, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[flow.jobs[-1].uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-46.8613738)
+    task_doc = responses[flow.jobs[-1].uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-46.8613738)
 
 
 def test_mp_gga_double_relax_static(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
-    from atomate2.vasp.flows.mp import MPGGADoubleRelaxStaticMaker
-
     # map from job name to directory containing reference output files
     pre_relax_dir = "Si_mp_gga_relax/GGA_Relax_1"
     ref_paths = {
@@ -106,17 +96,12 @@ def test_mp_gga_double_relax_static(mock_vasp, clean_dir, vasp_test_dir):
     responses = run_locally(flow, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[flow.jobs[-1].uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-10.84060922)
+    task_doc = responses[flow.jobs[-1].uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-10.84060922)
 
 
 def test_mp_gga_double_relax(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
-    from atomate2.vasp.flows.mp import MPGGADoubleRelaxMaker
-
     # map from job name to directory containing reference output files
     pre_relax_dir = "Si_mp_gga_relax/GGA_Relax_1"
     ref_paths = {
@@ -134,6 +119,6 @@ def test_mp_gga_double_relax(mock_vasp, clean_dir, vasp_test_dir):
     responses = run_locally(flow, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[flow.jobs[-1].uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-10.84145656)
+    task_doc = responses[flow.jobs[-1].uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-10.84145656)
