@@ -1,4 +1,6 @@
 import pytest
+from emmet.core.tasks import TaskDoc
+from jobflow import run_locally
 from pymatgen.core import Structure
 
 from atomate2.vasp.jobs.mp import (
@@ -42,9 +44,6 @@ def test_mp_meta_gga_relax_maker_default_values():
 
 
 def test_mp_meta_gga_static_maker(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
     # map from job name to directory containing reference output files
     ref_paths = {
         "MP meta-GGA static": "Si_mp_meta_gga_relax/r2scan_final_static",
@@ -58,7 +57,6 @@ def test_mp_meta_gga_static_maker(mock_vasp, clean_dir, vasp_test_dir):
 
     mock_vasp(ref_paths, fake_run_vasp_kwargs)
 
-    # generate flow
     job = MPMetaGGAStaticMaker(
         input_set_generator=MPMetaGGARelaxSetGenerator(auto_kspacing=0.8249)
     ).make(si_struct)
@@ -67,15 +65,12 @@ def test_mp_meta_gga_static_maker(mock_vasp, clean_dir, vasp_test_dir):
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[job.uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-46.8613738)
+    task_doc = responses[job.uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-46.8613738)
 
 
 def test_mp_meta_gga_relax_maker(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
     # map from job name to directory containing reference output files
     ref_paths = {
         "MP meta-GGA relax": "Si_mp_meta_gga_relax/r2scan_relax",
@@ -91,7 +86,6 @@ def test_mp_meta_gga_relax_maker(mock_vasp, clean_dir, vasp_test_dir):
 
     mock_vasp(ref_paths, fake_run_vasp_kwargs)
 
-    # generate flow
     job = MPMetaGGARelaxMaker(
         input_set_generator=MPMetaGGARelaxSetGenerator(auto_kspacing=0.4786)
     ).make(si_struct)
@@ -100,15 +94,12 @@ def test_mp_meta_gga_relax_maker(mock_vasp, clean_dir, vasp_test_dir):
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[job.uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-46.86703814)
+    task_doc = responses[job.uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-46.86703814)
 
 
 def test_mp_gga_relax_maker(mock_vasp, clean_dir, vasp_test_dir):
-    from emmet.core.tasks import TaskDoc
-    from jobflow import run_locally
-
     # map from job name to directory containing reference output files
     ref_paths = {
         "MP GGA relax": "Si_mp_gga_relax/GGA_Relax_1",
@@ -124,13 +115,12 @@ def test_mp_gga_relax_maker(mock_vasp, clean_dir, vasp_test_dir):
 
     mock_vasp(ref_paths, fake_run_vasp_kwargs)
 
-    # generate flow
     job = MPGGARelaxMaker().make(si_struct)
 
     # ensure flow runs successfully
     responses = run_locally(job, create_folders=True, ensure_success=True)
 
     # validate output
-    output = responses[job.uuid][1].output
-    assert isinstance(output, TaskDoc)
-    assert output.output.energy == pytest.approx(-10.84140641)
+    task_doc = responses[job.uuid][1].output
+    assert isinstance(task_doc, TaskDoc)
+    assert task_doc.output.energy == pytest.approx(-10.84140641)
