@@ -1,6 +1,7 @@
-def test_structure_optimization(vasp_test_dir):
-    from atomate2.vasp.drones import VaspDrone
+from atomate2.vasp.drones import VaspDrone
 
+
+def test_structure_optimization(vasp_test_dir):
     drone = VaspDrone()
     doc = drone.assimilate(vasp_test_dir / "Si_band_structure" / "static" / "outputs")
     assert doc
