@@ -1,4 +1,8 @@
+from pathlib import Path
+
 import pytest
+
+from atomate2.vasp.files import copy_vasp_outputs, get_largest_relax_extension
 
 
 @pytest.mark.parametrize(
@@ -10,10 +14,6 @@ import pytest
     ],
 )
 def test_copy_vasp_outputs_static(vasp_test_dir, tmp_dir, copy_kwargs, files):
-    from pathlib import Path
-
-    from atomate2.vasp.files import copy_vasp_outputs
-
     path = vasp_test_dir / "Si_band_structure" / "static" / "outputs"
     copy_vasp_outputs(src_dir=path, **copy_kwargs)
 
@@ -30,10 +30,6 @@ def test_copy_vasp_outputs_static(vasp_test_dir, tmp_dir, copy_kwargs, files):
     ],
 )
 def test_copy_vasp_outputs_double(vasp_test_dir, tmp_dir, copy_kwargs, files):
-    from pathlib import Path
-
-    from atomate2.vasp.files import copy_vasp_outputs
-
     path = vasp_test_dir / "Si_old_double_relax" / "outputs"
     copy_vasp_outputs(src_dir=path, **copy_kwargs)
 
@@ -42,8 +38,6 @@ def test_copy_vasp_outputs_double(vasp_test_dir, tmp_dir, copy_kwargs, files):
 
 
 def test_get_largest_relax_extension(vasp_test_dir):
-    from atomate2.vasp.files import get_largest_relax_extension
-
     path = vasp_test_dir / "Si_old_double_relax" / "outputs"
     extension = get_largest_relax_extension(directory=path)
     assert extension == ".relax2"
